Refactor webhook component
